cleaning up functions
