+++ generate-video-text.py
@@ -1,9 +1,4 @@
-import html
 import os
-import json
-import re
-import httpx  # type: ignore
-from bs4 import BeautifulSoup  # type: ignore
 from dotenv import load_dotenv  # type: ignore
 
 from utils.file import (
@@ -17,80 +12,13 @@ from utils.constants import TEMPLATE_DIR
 from utils.helpers import print_error, print_info, print_success
 from utils.media import get_video_upload_date
 from utils.format_helper import format_passage_ref
-from utils.studylight_api_helper import format_passage_ref_for_studylight
+from utils.studylight_api_helper import (
+    extract_and_format_text_from_studylight_html,
+    fetch_passage_via_studylight,
+    format_passage_ref_for_studylight,
+)
 from utils.time import get_formatted_date
 from utils.ui import confirm_parameters
-
-
-def fetch_passage_via_studylight(formatted_passage: str) -> object:
-    """
-    Fetches Bible passage data from StudyLight API.
-
-    Args:
-      formatted_passage (str): The formatted passage to fetch.
-
-    Returns:
-      object: The fetched data.
-
-    Raises:
-      ValueError: If AUTH_ID or AUTH_CODE environment variables are not set.
-    """
-    load_dotenv()
-
-    auth_id = os.getenv("STUDYLIGHT_AUTH_ID")
-    auth_code = os.getenv("STUDYLIGHT_AUTH_CODE")
-
-    if not auth_id or not auth_code:
-        raise ValueError("AUTH_ID and AUTH_CODE environment variables must be set.")
-
-    url = f"https://beta.studylight.org/sandbox/bible_verse.cgi?auth-id={auth_id}&auth-code={auth_code}&lang=eng&instructions=no&vsrefs=no&trans=n84&passage={formatted_passage}"
-
-    response = httpx.get(url)
-    if response.status_code == 200:
-        data = response.json()
-    else:
-        print("Failed to fetch data")
-
-    return data
-
-
-def extract_and_format_text_from_studylight_html(json_data) -> str:
-    """
-    Extracts and formats text from HTML content embedded within JSON or dict data,
-    ensuring that all text is processed by removing undesired HTML tags like <p>, <br>,
-    and heading tags (<h1>, <h2>, <h3>, etc.). Processes the text line by line to maintain
-    narrative flow as per each original line in the HTML, and prepends each line with
-    5 white spaces.
-
-    Args:
-        json_data (str or dict): JSON string or dict containing the HTML content.
-
-    Returns:
-        str: The cleaned, extracted text content, formatted to maintain narrative flow
-        with each line beginning with 5 white spaces.
-    """
-    if isinstance(json_data, dict):
-        data = json_data
-    else:
-        data = json.loads(json_data)
-
-    html_content = data["text"]
-    html_content = re.sub(r"<h[1-6].*?>.*?</h[1-6]>", "", html_content, flags=re.DOTALL)
-    html_content = re.sub(
-        r"<p[^>]*>|<br\s*/?>", "\n", html_content, flags=re.IGNORECASE
-    )
-    html_content = re.sub(r"</p>|<span[^>]*>|</span>", "", html_content)
-
-    lines = html_content.split("\n")
-    formatted_text = []
-
-    for line in lines:
-        line = html.unescape(line).strip()
-        if line:  # Ensure the line is not just whitespace
-            formatted_text.append("     " + line)
-
-    full_text = "\n".join(formatted_text)
-    return full_text
 
 
 def generate_title_from_template(
