+++ requirements.txt
@@ -12,7 +12,6 @@ requests==2.31.0
 tqdm==4.66.1
 urllib3==2.1.0
 yt-dlp==2024.4.9
-beautifulsoup4==4.11.1
 lxml==4.9.1
 python-dotenv==1.0.1
 httpx==0.27.0
